# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TAMAZIRT API Documentation",
        "description": "API for TAMAZIRT",
        "contact": {
            "email": "support@tamazirt.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "TAMAZIRT API Server"
        }
    ],
    "paths": {
...
        "/api/navigator-applications": {
            "get": {
                "tags": [
                    "Navigator Applications"
                ],
                "summary": "Get all navigator applications (Admin only)",
                "description": "Retrieves a list of all submitted applications. Requires 'admin' role.",
                "operationId": "adminIndexApplications",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "full_name": {
                                                        "type": "string",
                                                        "example": "Omar Benali"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized: admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Navigator Applications"
                ],
                "summary": "Submit a new navigator application (Public)",
                "description": "Allows any user to submit an application to become a navigator.",
                "operationId": "submitApplication",
                "requestBody": {
                    "description": "Application details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "phone_number"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Omar Benali"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "omar.benali@example.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+212600000000"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Marrakesh",
                                        "nullable": true
                                    },
                                    "motivation": {
                                        "type": "string",
                                        "example": "I want to share my knowledge of the city.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Application submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Application submitted successfully!"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": []
            }
        },
        "/api/navigator-applications/{id}/status": {
            "patch": {
                "tags": [
                    "Navigator Applications"
                ],
                "summary": "Update application status (Admin only)",
                "description": "Changes the status of a specific application (pending, approved, or rejected). Requires 'admin' role.",
                "operationId": "adminUpdateApplicationStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the application to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "approved",
                                            "rejected"
                                        ],
                                        "example": "approved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Application status updated successfully."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized: admin access required"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/navigator-applications/{id}": {
            "delete": {
                "tags": [
                    "Navigator Applications"
                ],
                "summary": "Delete an application (Admin only)",
                "description": "Deletes a specific navigator application. Requires 'admin' role.",
                "operationId": "adminDeleteApplication",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the application to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Application deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized: admin access required"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Bearer Token Authentication",
                "in": "header",
                "bearerFormat": "Sanctum Token",
                "scheme": "Bearer"
            }
        }
    },
...
}